Moved the new daemon code into the proper namespace
