+++ src/Marten.AsyncDaemon.Testing/NewDaemon/SubscriptionAgentTests.cs
@@ -2,7 +2,6 @@ using System.Threading;
 using System.Threading.Tasks;
 using Marten.Events;
 using Marten.Events.Daemon;
-using Marten.Events.Daemon.New;
 using Marten.Events.Projections;
 using Marten.Storage;
 using Microsoft.Extensions.Logging.Abstractions;
