+++ src/Marten.AsyncDaemon.Testing/build_aggregate_projection.cs
@@ -427,7 +427,7 @@ public class build_aggregate_projection: DaemonContext
             Assert.Equal("Foo", foo.Name);
         }
 
-        var daemon = theStore.BuildProjectionDaemon();
+        var daemon = await theStore.BuildProjectionDaemonAsync();
 
         await daemon.RebuildProjection("Foo", CancellationToken.None);
