+++ src/Marten/Events/Daemon/IDeadLetterQueue.cs
@@ -1,4 +1,4 @@
-namespace Marten.Events.Daemon.New;
+namespace Marten.Events.Daemon;
 
 public interface IDaemonRuntime
 {
