+++ src/Marten/Events/Daemon/IEventLoader.cs
@@ -1,7 +1,7 @@
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Marten.Events.Daemon.New;
+namespace Marten.Events.Daemon;
 
 public interface IEventLoader
 {
