+++ src/Marten/Events/Daemon/NulloDaemonRuntime.cs
@@ -1,4 +1,4 @@
-namespace Marten.Events.Daemon.New;
+namespace Marten.Events.Daemon;
 
 public class NulloDaemonRuntime: IDaemonRuntime
 {
